requirements.txt
# Django==4.2.16
# google-generativeai>=0.7.0
# google-genai==1.10.0
# Pillow==10.4.0
# channels==4.1.0
# daphne==4.1.2
# django-celery-beat==2.7.0
# celery==5.4.0
# redis==5.1.1
# requests==2.32.3
# python-dotenv==1.0.1
# opencv-python-headless==4.10.0.84
# mediapipe==0.10.14
# numpy==1.26.4
# sendgrid==6.11.0

# Django==4.2.16
# Pillow==10.4.0
# channels==4.1.0
# daphne==4.1.2
# django-celery-beat==2.7.0
# celery==5.4.0
# redis==5.1.1
# requests==2.32.3
# python-dotenv==1.0.1
# opencv-python-headless==4.10.0.84
# mediapipe==0.10.14
# numpy==1.26.4
# sendgrid==6.11.0

Django==4.2.16
daphne==4.2.1
channels==4.1.0

celery==5.4.0
django-celery-beat==2.7.0
redis==5.1.1

requests==2.32.3
python-dotenv==1.0.1

Pillow==10.4.0
numpy==1.26.4

opencv-python-headless==4.10.0.84
mediapipe==0.10.14

sendgrid==6.11.0

# IMPORTANT: removed problematic packages:
# any system-level linux bindings ❌
